+++ src/client.ts
@@ -102,6 +102,7 @@ export type {
     LambderI18nInstance,
     LambderI18nTranslator,
     LambderI18nExtractParams,
+    LambderI18nDictionaryLoader,
     LambderI18nCodes,
     LambderI18nKeys,
     LambderI18nTranslatorFor,